courses/databases-ii/practices/Mongo-Redis/pyproject.toml
[project]
name = "mongo-redis"
version = "0.1.0"
description = "This is an example using fastapi, redis and mongo db."
authors = [
    {name = "Carlos Andres Sierra",email = "cavirguezs@udistrital.edu.co"}
]
license = {text = "GNU/GPL 3.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "pymongo (>=4.15.4,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
